refactor: change entity and database identifiers from int to long

+++ src/Backend/GOtica.Domain/Entities/UserOpticalStore.cs
@@ -6,9 +6,9 @@ public class UserOpticalStore
     public string Role { get; set; } = string.Empty;
     public bool IsActive { get; set; }
 
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User User { get; set; } = default!;
 
-    public int OpticalStoreId { get; set; }
+    public long OpticalStoreId { get; set; }
     public OpticalStore OpticalStore { get; set; } = default!;
 }
